Titles, Styles and formatting
Title added. Changed a bit of style, table totals now display currency.

Added sqlquery.txt:sql query, duh.

Added selector:this one's going to be on a upper level than reports and
it will be renamed "main.jsx" and the actual "main.jsx" will be renamed
to "reports.jsx"

Added "estructura.p.js": which is the structure for products